Split up the monolithic System.h file
`System.h` contained way too many disparate parts. Separated hardware
abstractions from each other to make the libraries easier to read and
understand. Some more work still is to be done on UART libraries and
separating out a console from a UART interface.

In addition to this, future commits will work on cleaning up the
abstractions for the display hardware and, eventually, the display
software abstractions as well. The focus of these commits is to reduce
the number of monolithic classes and decrease average function size.
